+++ src/sgs_tools/io/um.py
@@ -5,6 +5,8 @@ from typing import Iterable
 import numpy as np
 import xarray as xr
 
+from sgs_tools.geometry.staggered_grid import interpolate_to_grid
+
 base_fields_dict = {
     "U_COMPNT_OF_WIND_AFTER_TIMESTEP": "u",
     "V_COMPNT_OF_WIND_AFTER_TIMESTEP": "v",
@@ -251,3 +253,40 @@ def compose_diagnostic_tensor(ds: xr.Dataset) -> xr.Dataset:
     ds["Diag_ij"] = diag_ij
     ds = ds.drop_vars(["diag11", "diag22", "diag33", "diag12", "diag13", "diag23"])
     return ds
+
+
+def data_ingest_UM_on_single_grid(
+    fname_pattern: Path,
+    res: float,
+    required_fields: list[str] = ["u", "v", "w", "theta"],
+):
+    """read and pre-process UM data
+
+    :param dir_path: directory containing
+    :param fname_pattern: UM NetCDF diagnostic file(s) to read. will be interpreted as a glob pattern. (should belong to the same simulation)
+    :param res: horizontal resolution (will use to overwrite horizontal coordinates). **NB** works for ideal simulations
+    :parm  required_fields: list of fields to read and pre-process. Defaults to ['u', 'v', 'w', 'theta']
+    """
+    # all the fields we will need for the Cs calculations
+    simulation = read_stash_files(fname_pattern)
+    # parse UM stash codes into variable names
+    simulation = rename_variables(simulation)
+
+    # rename to sgs_tools naming convention
+    simulation = standardize_varnames(simulation)
+
+    # restrict to interesting fields and rename to simple names
+    simulation = restrict_ds(simulation, fields=required_fields)
+
+    # unify coordinatesh
+    simulation = unify_coords(simulation, res=res)
+
+    # interpolate all vars to a cell-centred grid
+    centre_dims = ["x_centre", "y_centre", "z_theta"]
+    simulation = interpolate_to_grid(simulation, centre_dims, drop_coords=True)
+    # rename spatial dimensions to 'xyz'
+    simple_dims = ["x", "y", "z"]
+    dim_names = {d_new: d_old for d_new, d_old in zip(centre_dims, simple_dims)}
+    simulation = simulation.rename(dim_names)
+
+    return simulation
